Force format of x axis ticks in authorship plots
It's not the best solution to this issue, as on some plots we'll now see
the years several times. Another option would be to have %b %Y - but
again, on ietf-mls plotly is hallucinating dates - showing the 1st of
June as the last day of May, at 23:59. I consider I cannot fix that,
only attenuate it.

will-fix: #6